synthesis/requirements.txt
scipy
numpy
fastdtw
pyworld
pysptk
librosa
h5py
torch==1.10.1
torchaudio==0.10.1
torchvision==0.11.2
python-git
joblib
PyYAML
tqdm
wandb
jupyter
fairseq
unidecode
inflect
contractions
matplotlib
seaborn
